More AngelScript Unit Tests
* Some code cleanup
* Fixed issues calling functions with ezGameObject arguments
* Fixed missing function declarations on component types in as.predefined
* Added function overloads for functions that return ezVariant types
* Added messaging unit tests
* Added ezSpatial and physics unit tests
* Added unit tests for CVar access
* Added unit tests for arrays
* Added unit tests for instantiating prefabs
