+++ Code/EnginePlugins/AngelScriptPlugin/Runtime/AsEngineSingleton.h
@@ -70,11 +70,10 @@ private:
   void Register_ReflectedType(const ezRTTI* pBaseType, bool bCreatable);
   void Register_ReflectedTypes();
   void RegisterTypeFunctions(const char* szTypeName, const ezRTTI* pRtti, bool bIsInherited);
-  void RegisterTypeProperties(const char* szTypeName, const ezRTTI* pRtti, bool bIsInherited);
+  void RegisterSingleGenericFunction(const char* szFuncName, const char* szTypeName, const ezAbstractFunctionProperty* const pFunc, const ezScriptableFunctionAttribute* pFuncAttr, bool bIsInherited, const ezRTTI* pReturnType);
   void RegisterGenericFunction(const char* szTypeName, const ezAbstractFunctionProperty* const pFunc, const ezScriptableFunctionAttribute* pFuncAttr, bool bIsInherited);
   bool AppendType(ezStringBuilder& decl, const ezRTTI* pRtti, const ezScriptableFunctionAttribute* pFuncAttr, ezUInt32 uiArg, bool& inout_VarArgs);
   bool AppendFuncArgs(ezStringBuilder& decl, const ezAbstractFunctionProperty* pFunc, const ezScriptableFunctionAttribute* pFuncAttr, ezUInt32 uiArg, bool& inout_VarArgs);
-  ezString Register_EnumType(const ezRTTI* pEnumType);
   void Register_ExtraComponentFuncs();
