+++ Code/UnitTests/GameEngineTest/AngelScriptTest/AngelScriptTest.h
@@ -29,6 +29,8 @@ public:
     Strings,
     Messaging,
     GameObject,
+    Physics,
+    Misc,
   };
 
 private:
