+++ Data/Samples/SampleGame/Editor/AngelScript/GlobalFunctions.txt
@@ -38,6 +38,19 @@ GetScriptOwnerObject
 GetScriptOwnerWorld
 GetStringValue
 GetType
+GetValue_asAngle
+GetValue_asBool
+GetValue_asColor
+GetValue_asComponentHandle
+GetValue_asFloat
+GetValue_asGameObjectHandle
+GetValue_asInt32
+GetValue_asQuat
+GetValue_asString
+GetValue_asTime
+GetValue_asVec2
+GetValue_asVec3
+GetValue_asVec4
 Hours
 Info
 IsEqual
