+++ Data/Samples/SampleGame/Editor/AngelScript/Methods.txt
@@ -77,6 +77,19 @@ GetDarker
 GetDegree
 GetDistanceTo
 GetElementCount
+GetEntryValue_asAngle
+GetEntryValue_asBool
+GetEntryValue_asColor
+GetEntryValue_asComponentHandle
+GetEntryValue_asFloat
+GetEntryValue_asGameObjectHandle
+GetEntryValue_asInt32
+GetEntryValue_asQuat
+GetEntryValue_asString
+GetEntryValue_asTime
+GetEntryValue_asVec2
+GetEntryValue_asVec3
+GetEntryValue_asVec4
 GetFileDirectory
 GetFileExtension
 GetFileName
@@ -124,6 +137,19 @@ GetReflectedVector
 GetRootedPathRootName
 GetRotationAxisAndAngle
 GetSaturation
+GetScriptVariable_asAngle
+GetScriptVariable_asBool
+GetScriptVariable_asColor
+GetScriptVariable_asComponentHandle
+GetScriptVariable_asFloat
+GetScriptVariable_asGameObjectHandle
+GetScriptVariable_asInt32
+GetScriptVariable_asQuat
+GetScriptVariable_asString
+GetScriptVariable_asTime
+GetScriptVariable_asVec2
+GetScriptVariable_asVec3
+GetScriptVariable_asVec4
 GetSeconds
 GetShrunk
 GetSpeed
@@ -133,6 +159,19 @@ GetSubString
 GetTeamID
 GetTimeDiff
 GetUniqueID
+GetValue_asAngle
+GetValue_asBool
+GetValue_asColor
+GetValue_asComponentHandle
+GetValue_asFloat
+GetValue_asGameObjectHandle
+GetValue_asInt32
+GetValue_asQuat
+GetValue_asString
+GetValue_asTime
+GetValue_asVec2
+GetValue_asVec3
+GetValue_asVec4
 GetVectorPart
 GetView
 GetWorld
