+++ Data/Samples/SampleGame/Editor/AngelScript/NotRegisteredDecls.txt
@@ -1,2 +1,1 @@
 ezBlackboard
-ezVariant
