+++ Data/UnitTests/GameEngineTest/AngelScript/Editor/AngelScript/Enums.txt
@@ -19,6 +19,7 @@ Center
 Character
 ClearEntireBlackboard
 Column
+Cone
 Continuing
 Culling
 CustomPose
@@ -43,18 +44,31 @@ DeviceID8
 DeviceID9
 Directional
 Disabled
+DrivePosition
+DriveVelocity
 Dynamic
 EditorView
 EnsureEntriesExist
 EveryFrame
 ExtremeShockwave
+Fixed
+FixedCornerBottomLeft
+FixedCornerBottomRight
+FixedCornerTopLeft
+FixedCornerTopRight
+FixedEdgeBottom
+FixedEdgeLeft
+FixedEdgeRight
+FixedEdgeTop
 FullRotation
 FullUpdate
 GentleBreeze
 Global
 Grip
 HMD
+HardLimit
 Ignore
+ImpactReactions
 Left
 LeftController
 LightBreeze
@@ -84,6 +98,8 @@ NegativeY
 NegativeZ
 Never
 NextFrame
+NoDrive
+NoLimit
 None
 Once
 OnlyPosition
@@ -92,6 +108,7 @@ OrthoFixedWidth
 Pause
 PerspectiveFixedFovX
 PerspectiveFixedFovY
+Point
 PositiveX
 PositiveY
 PositiveZ
@@ -103,6 +120,7 @@ PreserveLocal
 PressAndRelease
 PressOnly
 PressReleaseAndDown
+PrintState
 QuadPosX
 Query
 Ragdoll
@@ -115,12 +133,17 @@ ResetEntryValues
 RestPose
 Right
 RightController
+RollXReactions
+RollYReactions
+RollZReactions
 Rope
 Seated
+SemiFixed
 SendMoveCharacterMsg
 SetAlpha
 SetRGB
 SetRGBA
+SlideReactions
 SpiralStairs
 StairsNegX
 StairsNegY
@@ -132,13 +155,23 @@ Storm
 StrongBreeze
 StrongShockwave
 Subtractive
+SwingTwist
 Thumbnail
 TopCenter
 TopLeft
 TopRight
 Transparent
 Trigger
+Variable
+VisCasts
+VisContacts
+VisFootCheck
+VisGroundContact
+VisShape
 VisualizePath
 VisualizeUpDir
 Vortex
 WeakShockwave
+WithBindPose
+WithCurrentMeshPose
+WithNextAnimPose
