+++ Data/UnitTests/GameEngineTest/AngelScript/Editor/AngelScript/Methods.txt
@@ -1,5 +1,10 @@
 Abs
+AddAngularForce
+AddAngularImpulse
 AddChild
+AddInitialImpulse
+AddLinearForce
+AddLinearImpulse
 Append
 AppendFormat
 AppendPath
@@ -8,6 +13,7 @@ ApplyHdrExposureValue
 AsFloatInSeconds
 Assign
 Bool
+BreakObjectGrab
 BroadcastEventMsg
 CalcAverageRGB
 CanTriggerManualSpawn
@@ -43,6 +49,7 @@ DoubleVariance
 DoubleVarianceAroundZero
 DoubleZeroToOneExclusive
 DoubleZeroToOneInclusive
+DropGrabbedObject
 EndsWith
 EndsWith_NoCase
 ExpandAndGetRef
@@ -77,6 +84,19 @@ GetDarker
 GetDegree
 GetDistanceTo
 GetElementCount
+GetEntryValue_asAngle
+GetEntryValue_asBool
+GetEntryValue_asColor
+GetEntryValue_asComponentHandle
+GetEntryValue_asFloat
+GetEntryValue_asGameObjectHandle
+GetEntryValue_asInt32
+GetEntryValue_asQuat
+GetEntryValue_asString
+GetEntryValue_asTime
+GetEntryValue_asVec2
+GetEntryValue_asVec3
+GetEntryValue_asVec4
 GetFileDirectory
 GetFileExtension
 GetFileName
@@ -114,6 +134,7 @@ GetNanoseconds
 GetNegated
 GetNormalized
 GetNormalizedRange
+GetObjectFilterID
 GetOrthogonalVector
 GetOwner
 GetParent
@@ -124,6 +145,19 @@ GetReflectedVector
 GetRootedPathRootName
 GetRotationAxisAndAngle
 GetSaturation
+GetScriptVariable_asAngle
+GetScriptVariable_asBool
+GetScriptVariable_asColor
+GetScriptVariable_asComponentHandle
+GetScriptVariable_asFloat
+GetScriptVariable_asGameObjectHandle
+GetScriptVariable_asInt32
+GetScriptVariable_asQuat
+GetScriptVariable_asString
+GetScriptVariable_asTime
+GetScriptVariable_asVec2
+GetScriptVariable_asVec3
+GetScriptVariable_asVec4
 GetSeconds
 GetShrunk
 GetSpeed
@@ -133,13 +167,28 @@ GetSubString
 GetTeamID
 GetTimeDiff
 GetUniqueID
+GetValue_asAngle
+GetValue_asBool
+GetValue_asColor
+GetValue_asComponentHandle
+GetValue_asFloat
+GetValue_asGameObjectHandle
+GetValue_asInt32
+GetValue_asQuat
+GetValue_asString
+GetValue_asTime
+GetValue_asVec2
+GetValue_asVec3
+GetValue_asVec4
 GetVectorPart
 GetView
 GetWorld
+GrabNearbyObject
 HasAnyExtension
 HasExtension
 HasMirrorScaling
 HasName
+HasObjectGrabbed
 HasTag
 IndexOf
 IndexOfByRef
@@ -170,6 +219,7 @@ IsEqual_NoCase
 IsIdentical
 IsIdenticalRGB
 IsIdenticalRGBA
+IsInAir
 IsInvalidated
 IsNaN
 IsNegative
@@ -177,6 +227,8 @@ IsNormalized
 IsPositive
 IsRelativePath
 IsRootedPath
+IsSlidingOnGround
+IsStandingOnGround
 IsStatic
 IsTouchingGround
 IsValid
@@ -208,6 +260,7 @@ PushBack
 RawMove
 ReconstructFromMat3
 ReconstructFromMat4
+RecreatePhysicsShapes
 RemoveAtAndCopy
 RemoveFileExtension
 ReplaceAll
@@ -233,6 +286,8 @@ SetGlobalRotation
 SetGlobalScaling
 SetGlobalTransform
 SetIdentity
+SetInitialImpulse
+SetJointTypeOverride
 SetLocalPosition
 SetLocalRotation
 SetLocalScaling
@@ -261,6 +316,7 @@ StartTransitionWithOffsetTo
 StartsWith
 StartsWith_NoCase
 TeleportCharacter
+ThrowGrabbedObject
 ToLinearFloat
 ToLower
 ToUpper
