+++ Data/UnitTests/GameEngineTest/AngelScript/Editor/AngelScript/Properties.txt
@@ -1,8 +1,16 @@
 Acceleration
 ActivationDelay
 Active
+AirFriction
+AirSpeed
 AliceBlue
+AllowSleeping
+Anchor1
+Anchor1Constraint
+Anchor2
+Anchor2Constraint
 Angle
+AngularDamping
 AnimGraph
 Animation
 AnimationClip
@@ -16,10 +24,12 @@ Aqua
 Aquamarine
 AspectRatio
 AttachAsChild
+AttachTo
 Axis
 AxisDeviation
 Azure
 Beige
+BendStiffness
 Bisque
 Black
 BlackboardName
@@ -32,31 +42,48 @@ BottomColor
 BoundsOffset
 BoundsRadius
 BoxProjection
+BreakDistance
+BreakForce
+BreakTorque
 Brown
 BurlyWood
 BurstDuration
 CadetBlue
 CameraRenderPipeline
 CaptureOffset
+CastRadius
 CastShadows
+CenterAngularVelocity
+CenterOfMass
+CenterPosition
+CenterVelocity
+CharacterGravity
 Chartreuse
+ChildActor
+ChildActorAnchor
 Chocolate
 Closed
 Collection
 CollisionLayer
+CollisionMesh
 Color
 Comment
+ConeAngle
 ConstantBias
+ContinuousCollisionDetection
 Coral
 CornflowerBlue
 Cornsilk
 Crimson
 Crouch
+CrouchHeight
 CubeMap
 Curvature
+CustomCenterOfMass
 CustomData
 Cyan
 Damage
+Damping
 DarkBlue
 DarkCyan
 DarkGoldenRod
@@ -81,6 +108,7 @@ Data1
 Data2
 Data3
 DeactivationDelay
+DebugFlags
 DebugShowPoints
 Deceleration
 DeepPink
@@ -97,6 +125,9 @@ Dimensions
 Directionality
 Distance
 DodgerBlue
+DriveMode
+DriveStrength
+DriveTargetValue
 Duration
 EditBones
 EditorShortcut
@@ -109,18 +140,23 @@ Extents
 FOV
 FadeOutDuration
 FadeOutStart
+FallbackWalkSurface
 Falloff
 FalloffExponent
 FarDistance
 FarPlane
 FireBrick
+FixedFrameRate
 Flags
 FloralWhite
 FollowMode
+FootRadius
 Force
 ForestGreen
 ForwardToBlackboard
 ForwardVector
+Frequency
+Friction
 Fuchsia
 Gainsboro
 GameObject
@@ -130,16 +166,19 @@ GlobalPosition
 Gold
 GoldenRod
 GotGrabbed
+GrabAnyObjectWithSize
 GrabbedBy
 GrabbedObjectToRelease
 Gradient
 Granularity
+GravityFactor
 Gray
 Green
 GreenYellow
 Grey
 HalfExtents
 HandleGlobalEvents
+HeadObject
 Height
 HeightFalloff
 HitObjectName
@@ -169,8 +208,10 @@ JointName
 JointObject
 JointStart
 Jump
+JumpImpulse
 KeyPressValue
 Khaki
+Kinematic
 Lavender
 LavenderBlush
 LawnGreen
@@ -195,6 +236,8 @@ LightSteelBlue
 LightYellow
 Lime
 LimeGreen
+LimitMode
+LinearDamping
 Linen
 LinkToLightShape
 LocalPosition
@@ -202,19 +245,30 @@ LocalRotation
 LocalScaling
 LocalUniformScaling
 LookAhead
+LowerLimit
+LowerTwistLimit
 Magenta
 MainColor
 MapNormalToGeometry
 Marker
 Maroon
+Mass
 Material
 MaterialSlot
+MaxBend
+MaxBodies
+MaxClimbingSlope
 MaxDeviation
 MaxDistance
+MaxGrabPointDistance
 MaxRotationZ
 MaxScreenSize
+MaxStepDown
+MaxStepUp
+MaxSubSteps
 MaxTilt
 MaxTiltZ
+MaxTwist
 MaxWindStrength
 MediumAquaMarine
 MediumBlue
@@ -230,6 +284,7 @@ Message
 MidAxis
 MidnightBlue
 MinDelay
+MinDistance
 MinShadowRange
 MinSpawnCount
 MinWindStrength
@@ -251,6 +306,7 @@ NewStateName
 NumCascades
 Object
 ObjectFilterID
+ObjectGravity
 OcclusionDepthOffset
 OcclusionSampleRadius
 OcclusionSampleSpread
@@ -258,6 +314,7 @@ OldLace
 OldStateName
 Olive
 OliveDrab
+OnContact
 OnFinishedAction
 Orange
 OrangeRed
@@ -268,17 +325,21 @@ OverlapRanges
 OverridePosition
 OverrideRotation
 OverrideScale
+OwnerVelocityScale
+PairCollision
 PaleGoldenRod
 PaleGreen
 PaleTurquoise
 PaleVioletRed
 PapayaWhip
+ParentActor
 PassThroughUnhandledEvents
 Path
 Payload
 PeachPuff
 PenumbraSize
 Peru
+Pieces
 Pink
 Play
 PlayerPrefab
@@ -291,8 +352,11 @@ PositiveFalloff
 PowderBlue
 Prefab
 PreloadCollection
+PresenceCollisionLayer
 ProjectionAxis
+PullSurfacesFromGraphicsMesh
 Purple
+QueryShapeOnly
 Radius
 RadiusFalloff
 RandomOffset
@@ -322,6 +386,7 @@ RootMotionMode
 RosyBrown
 RotateLeft
 RotateRight
+RotateSpeed
 Rotation
 RotationOffset
 RotationX
@@ -329,6 +394,7 @@ RotationY
 RotationZ
 RoyalBlue
 Run
+RunInteractionDistance
 Running
 SaddleBrown
 Salmon
@@ -338,10 +404,12 @@ Scale
 ScriptClass
 SeaGreen
 SeaShell
+SelfCollision
 SendEntryChangedMessage
 SetColorMode
 ShadowFadeOutRange
 Shape
+ShapeRadius
 ShowDebugInfo
 ShowDebugOverlay
 ShowDebugProbes
@@ -361,9 +429,11 @@ SizeVariance
 Skeleton
 SkyBlue
 SkyDistance
+Slack
 SlateBlue
 SlateGray
 SlateGrey
+SleepVelocityThreshold
 SlopeBias
 SlopedBottom
 SlopedTop
@@ -380,15 +450,25 @@ SpecularMultiplier
 Speed
 SphereProjection
 SplitModeWeight
+SpringDamping
 SpringGreen
+SpringStiffness
 StageSpace
+StandHeight
+StartAsleep
 StartDistance
+StartMode
 SteelBlue
+SteppingMode
+StiffnessFactor
 StrafeLeft
 StrafeRight
 Strength
 StrengthFactor
 Subdivide
+Surface
+SwingLimitY
+SwingLimitZ
 Tan
 Tangent1
 Tangent2
@@ -401,6 +481,7 @@ TestPosition
 TestVisibility
 Text
 Texture
+TextureScale
 Thickness
 Thistle
 TiltAmount
@@ -410,6 +491,7 @@ TopColor
 TransformSpace
 Translation
 TransparentShadows
+TriggerMessage
 TriggerState
 Turquoise
 Type
@@ -418,6 +500,9 @@ UVUnitsPerWorldUnit
 UpVector
 UpdateInterval
 UpdateRate
+UpdateThreshold
+UpperLimit
+UpperTwistLimit
 UsageHint
 UseAsOccluder
 UseColorTemperature
@@ -436,11 +521,17 @@ VisualizeSkeleton
 VisualizeSwingLimits
 VisualizeTwistLimits
 VolumeType
+WalkInteractionDistance
+WalkSpeedCrouching
+WalkSpeedRunning
+WalkSpeedStanding
+WalkSurfaceInteraction
 Weight
 Wheat
 White
 WhiteSmoke
 Width
+WindInfluence
 WrapAround
 Yellow
 YellowGreen
