+++ Data/UnitTests/GameEngineTest/AngelScript/Editor/AngelScript/Types.txt
@@ -57,6 +57,42 @@ ezInputMessageGranularity
 ezInstancedMeshComponent
 ezJointAttachmentComponent
 ezJointOverrideComponent
+ezJoltActorComponent
+ezJoltCharacterControllerComponent
+ezJoltCharacterDebugFlags
+ezJoltClothSheetComponent
+ezJoltClothSheetFlags
+ezJoltConeConstraintComponent
+ezJoltConstraintComponent
+ezJoltConstraintDriveMode
+ezJoltConstraintLimitMode
+ezJoltDefaultCharacterComponent
+ezJoltDistanceConstraintComponent
+ezJoltDynamicActorComponent
+ezJoltFixedConstraintComponent
+ezJoltGrabObjectComponent
+ezJoltHingeConstraintComponent
+ezJoltHitboxComponent
+ezJoltMsgDisconnectConstraints
+ezJoltPointConstraintComponent
+ezJoltQueryShapeActorComponent
+ezJoltRagdollComponent
+ezJoltRagdollStartMode
+ezJoltRopeAnchorConstraintMode
+ezJoltRopeComponent
+ezJoltSettingsComponent
+ezJoltShapeBoxComponent
+ezJoltShapeCapsuleComponent
+ezJoltShapeComponent
+ezJoltShapeConvexHullComponent
+ezJoltShapeCylinderComponent
+ezJoltShapeSphereComponent
+ezJoltSliderConstraintComponent
+ezJoltStaticActorComponent
+ezJoltSteppingMode
+ezJoltSwingTwistConstraintComponent
+ezJoltTriggerComponent
+ezJoltVisColMeshComponent
 ezLensFlareComponent
 ezLightComponent
 ezLocalBlackboardComponent
@@ -115,6 +151,7 @@ ezObjectMsgQueueType
 ezOccluderComponent
 ezOccluderType
 ezOnComponentFinishedAction
+ezOnJoltContact
 ezPathComponent
 ezPathComponentFlags
 ezPathNodeComponent
@@ -150,6 +187,7 @@ ezShapeIconComponent
 ezSimpleAnimationComponent
 ezSimpleWindComponent
 ezSkeletonComponent
+ezSkeletonJointType
 ezSkeletonPoseComponent
 ezSkeletonPoseMode
 ezSkyBoxComponent
