+++ Data/UnitTests/GameEngineTest/AngelScript/Editor/PluginSelection.ddl
@@ -4,3 +4,9 @@ PluginState
 	bool %Selected{true}
 	bool %LoadCopy{false}
 }
+PluginState
+{
+	string %ID{"Jolt"}
+	bool %Selected{true}
+	bool %LoadCopy{false}
+}
