+++ Data/UnitTests/GameEngineTest/AngelScript/Editor/SpatialDataCategoryEnum.txt
@@ -1,0 +1,1 @@
+Marker
