+++ Data/UnitTests/GameEngineTest/AngelScript/RuntimeConfigs/Plugins.ddl
@@ -3,3 +3,8 @@ Plugin
 	string %Path{"ezAngelScriptPlugin"}
 	bool %LoadCopy{false}
 }
+Plugin
+{
+	string %Path{"ezJoltPlugin"}
+	bool %LoadCopy{false}
+}
